Added BPC Score Comparison Screen Flow.

+++ app/src/main/java/com/patsurvey/nudge/model/response/DidiDetailList.kt
@@ -18,8 +18,8 @@ data class DidiDetailList(
     @SerializedName("score") var score: Double,
     @SerializedName("comment") var comment: String,
     @SerializedName("beneficiaryProcessStatus") var beneficiaryProcessStatus: List<BeneficiaryProcessStatusModel>,
-    @SerializedName("bpcScore") var bpcScore: Int,
+    @SerializedName("bpcScore") var bpcScore: Double,
     @SerializedName("bpcComment") var bpcComment: String,
-    @SerializedName("crpScore") var crpScore: Int,
+    @SerializedName("crpScore") var crpScore: Double,
     @SerializedName("crpComment") var crpComment: String
 )
